requirements.txt
# Core dependencies
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# HTTP and utilities
requests>=2.31.0
isodate>=0.6.1

# Environment variable management
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.0
black>=23.7.0
flake8>=6.1.0
types-requests>=2.31.0
black>=24.0.0

# Logging and monitoring (optional)
structlog>=23.1.0
python-json-logger>=2.0.7